+++ uniffi_bindgen/src/bindings/kotlin/gen_kotlin/mod.rs
@@ -16,11 +16,9 @@ use crate::MergeWith;
 
 mod callback_interface;
 mod compounds;
-mod decorator;
 mod enum_;
 mod error;
 mod function;
-mod generic;
 mod miscellany;
 mod object;
 mod primitives;
@@ -103,9 +101,6 @@ impl<'a> KotlinWrapper<'a> {
         )
         .chain(ci.iter_function_definitions().into_iter().map(|inner| {
             Box::new(function::KotlinFunction::new(inner, ci)) as Box<dyn CodeDeclaration>
-        }))
-        .chain(ci.iter_decorator_definitions().into_iter().map(|inner| {
-            Box::new(decorator::KotlinDecoratorObject::new(inner, ci)) as Box<dyn CodeDeclaration>
         }))
         .chain(ci.iter_object_definitions().into_iter().map(|inner| {
             Box::new(object::KotlinObject::new(inner, ci)) as Box<dyn CodeDeclaration>
@@ -202,13 +197,12 @@ impl KotlinCodeOracle {
 
             Type::Enum(id) => Box::new(enum_::EnumCodeType::new(id)),
             Type::Object(id) => Box::new(object::ObjectCodeType::new(id)),
-            Type::Generic => Box::new(generic::GenericCodeType::new()),
             Type::Record(id) => Box::new(record::RecordCodeType::new(id)),
             Type::Error(id) => Box::new(error::ErrorCodeType::new(id)),
             Type::CallbackInterface(id) => {
                 Box::new(callback_interface::CallbackInterfaceCodeType::new(id))
             }
-            Type::DecoratorObject(s) => Box::new(decorator::DecoratorObjectCodeType::new(s)),
+
             Type::Optional(ref inner) => {
                 let outer = type_.clone();
                 let inner = *inner.to_owned();
@@ -306,13 +300,6 @@ pub mod filters {
         Ok(codetype.type_label(&oracle()))
     }
 
-    pub fn type_t_kt(
-        codetype: &impl CodeType,
-        t: &dyn fmt::Display,
-    ) -> Result<String, askama::Error> {
-        Ok(codetype.type_t_label(&oracle(), &t.to_string()))
-    }
-
     pub fn canonical_name(codetype: &impl CodeType) -> Result<String, askama::Error> {
         Ok(codetype.canonical_name(&oracle()))
     }
@@ -385,53 +372,5 @@ pub mod filters {
     /// and is distinguished from an "Exception".
     pub fn exception_name_kt(nm: &dyn fmt::Display) -> Result<String, askama::Error> {
         Ok(oracle().error_name(nm))
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_type_t_label() -> Result<()> {
-        let oracle = KotlinCodeOracle;
-        assert_eq!(
-            "Int".to_string(),
-            oracle
-                .find(&Type::Int32)
-                .type_t_label(&oracle, "GenericType")
-        );
-        assert_eq!(
-            "GenericType".to_string(),
-            oracle
-                .find(&Type::Generic)
-                .type_t_label(&oracle, "GenericType")
-        );
-        assert_eq!(
-            "GenericType?".to_string(),
-            oracle
-                .find(&Type::Optional(Box::new(Type::Generic)))
-                .type_t_label(&oracle, "GenericType")
-        );
-        assert_eq!(
-            "List<GenericType>".to_string(),
-            oracle
-                .find(&Type::Sequence(Box::new(Type::Generic)))
-                .type_t_label(&oracle, "GenericType")
-        );
-        assert_eq!(
-            "Map<String, GenericType>".to_string(),
-            oracle
-                .find(&Type::Map(Box::new(Type::Generic)))
-                .type_t_label(&oracle, "GenericType")
-        );
-
-        assert_eq!(
-            "T?".to_string(),
-            oracle
-                .find(&Type::Optional(Box::new(Type::Generic)))
-                .type_t_label(&oracle, "T")
-        );
-        Ok(())
     }
 }