+++ uniffi_bindgen/src/bindings/kotlin/templates/macros.kt
@@ -50,12 +50,6 @@
     {%- endfor %}
 {%- endmacro %}
 
-{% macro arg_list_protocol(func) %}
-    {%- for arg in func.arguments() -%}
-        {{ arg.name()|var_name_kt }}: {{ arg|type_kt -}}
-        {%- if !loop.last %}, {% endif -%}
-    {%- endfor %}
-{%- endmacro %}
 {#-
 // Arglist as used in the _UniFFILib function declations.
 // Note unfiltered name but type_ffi filters.