+++ uniffi_bindgen/src/bindings/python/gen_python.rs
@@ -162,10 +162,6 @@ mod filters {
             ),
             Type::Wrapped { prim, .. } => coerce_py(nm, prim.as_ref())?,
             Type::External { .. } => panic!("should not be necessary to coerce External types"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 
@@ -199,10 +195,6 @@ mod filters {
             ),
             Type::Wrapped { prim, .. } => lower_py(nm, prim.as_ref())?,
             Type::External { .. } => panic!("should not be necessary to lower External types"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 
@@ -235,10 +227,6 @@ mod filters {
             ),
             Type::Wrapped { prim, .. } => lift_py(nm, prim.as_ref())?,
             Type::External { .. } => panic!("should not be necessary to lift External types"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 }