+++ uniffi_bindgen/src/bindings/ruby/gen_ruby/mod.rs
@@ -192,10 +192,6 @@ mod filters {
             }
             Type::External { .. } => panic!("No support for external types, yet"),
             Type::Wrapped { .. } => panic!("No support for wrapped types, yet"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 
@@ -229,10 +225,6 @@ mod filters {
             ),
             Type::External { .. } => panic!("No support for lowering external types, yet"),
             Type::Wrapped { .. } => panic!("No support for lowering wrapped types, yet"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 
@@ -265,10 +257,6 @@ mod filters {
             ),
             Type::External { .. } => panic!("No support for lifting external types, yet"),
             Type::Wrapped { .. } => panic!("No support for lifting wrapped types, yet"),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         })
     }
 }