+++ uniffi_bindgen/src/bindings/swift/gen_swift/mod.rs
@@ -311,10 +311,6 @@ impl SwiftCodeOracle {
                 name,
                 self.create_code_type(prim.as_ref().clone()),
             )),
-            Type::DecoratorObject(_) => {
-                unreachable!("Decorator objects should never cross the FFI")
-            }
-            Type::Generic => unreachable!("Generic types should never cross the FFI"),
         }
     }
 }