+++ uniffi_bindgen/src/interface/types/resolver.rs
@@ -183,7 +183,6 @@ pub(in super::super) fn resolve_builtin_type(name: &str) -> Option<Type> {
         "f64" => Some(Type::Float64),
         "timestamp" => Some(Type::Timestamp),
         "duration" => Some(Type::Duration),
-        "Any" => Some(Type::Generic),
         _ => None,
     }
 }